+++ theme-suite/branches/2.0/src/Partial/ArticleBodyPartial.php
@@ -32,9 +32,11 @@ class ArticleBodyPartial extends AbstractPartialDriver
 
             return parent::render();
         } elseif ($article = ($p = $this->get('post', null)) instanceof QueryPostContract ? $p : post::create($p)) {
-            $enabled = ($article instanceof ThemeSuiteQueryPost)
-                ? array_merge($article->getSingularComposing('enabled', []), $this->get('enabled', []))
-                : $this->get('enabled', []);
+            if ($article instanceof ThemeSuiteQueryPost) {
+                $enabled = array_merge($article->getSingularComposing('enabled', []), $this->get('enabled', []));
+            } else {
+                $enabled = $this->get('enabled', []);
+            }
 
             $this->set([
                 'article' => $article,