+++ theme-suite/branches/2.0/src/Partial/ArticleChildrenPartial.php
@@ -35,9 +35,11 @@ class ArticleChildrenPartial extends AbstractPartialDriver
                 return '';
             }
 
-            $enabled = ($article instanceof ThemeSuiteQueryPost)
-                ? array_merge($article->getSingularComposing('enabled', []), $this->get('enabled', []))
-                : $this->get('enabled', []);
+            if ($article instanceof ThemeSuiteQueryPost) {
+                $enabled = array_merge($article->getSingularComposing('enabled', []), $this->get('enabled', []));
+            } else  {
+                $enabled = $this->get('enabled', []);
+            }
 
             if (!$enabled['children']) {
                 return '';
@@ -53,6 +55,7 @@ class ArticleChildrenPartial extends AbstractPartialDriver
             $this->set([
                 'article' => $article,
                 'items'   => $items,
+                'enabled' => $enabled
             ]);
 
             return parent::render();